GUI code cleanup. Fixes #3354.
Lots of code style fixes:
 * type [*&]varname -> type[*&] varname
 * else
   if (...)
   ->
   else if (...)
 * Spaces around some ops.
 * i++ -> ++i.
 * switch-case style fixes.
 * Indentation fixes.
 * Removing some commented out code.
 * include header sorting
 * Changed multiple pointer var declarations to be one per line.
 * Removed strange spaces in some places.
 * Changed some include header guards to be consistent with the rest of the codebase.
 * Use UNUSED() instead of UNUSED2().

Some small code fixes:
 * Using .find() instead of .count() == 0.
 * !.empty() instead of .size() == 0.
 * Range-based for loops.
 * Making some member functions const by small changes.
   * Adds GetScrollBarPos(idx) const for this.
 * Using early returns/continues in some places.
 * Uses size_t for some loops in CList and COList.
 * Removes unused heading element (not attribute) from COList.
 * Use ENSURE in one case where some custom code did something similar.
 * Made some parameters const ptrs/refs.
 * Change removal loop in GUItext.cpp to erase-unique.
 * Made some static things const.
 * Allow iterating over children of IGUIObject with range-based for loops by
   exposing begin() and end() (rename from ChildrenIt{Begin,End}()) and use it.

Comments:
 * Comment COList.
 * Update a few comments.
 * Remove useless or duplicated comments.

git-svn-id: http://svn.wildfiregames.com/public/ps/trunk@16931 3db68df2-c116-0410-a063-a993310a9797

+++ source/gui/COList.h
@@ -17,23 +17,8 @@
 #ifndef INCLUDED_COLIST
 #define INCLUDED_COLIST
 
-//--------------------------------------------------------
-//  Includes / Compiler directives
-//--------------------------------------------------------
 #include "GUI.h"
 #include "CList.h"
-
-//--------------------------------------------------------
-//  Macros
-//--------------------------------------------------------
-
-//--------------------------------------------------------
-//  Types
-//--------------------------------------------------------
-
-//--------------------------------------------------------
-//  Declarations
-//--------------------------------------------------------
 
 struct ObjectDef
 {
@@ -45,8 +30,12 @@ struct ObjectDef
 };
 
 /**
- *  Todo : add description
+ * Multi-column list. One row can be selected by the user.
+ * Individual cells are clipped if the contained text is too long.
  *
+ * The list can be sorted dynamically by JS code when a
+ * heading is ckicked.
+ * A scroll-bar will appear when needed.
  */
 class COList : public CList
 {
@@ -57,20 +46,19 @@ public:
 
 protected:
 	void SetupText();
-	void HandleMessage(SGUIMessage &Message);
+	void HandleMessage(SGUIMessage& Message);
 
 	/**
 	 * Handle the \<item\> tag.
 	 */
 	virtual bool HandleAdditionalChildren(const XMBElement& child, CXeromyces* pFile);
 
-	void DrawList(const int &selected, const CStr& _sprite,
-					const CStr& _sprite_selected, const CStr& _textcolor);
+	void DrawList(const int& selected, const CStr& _sprite, const CStr& _sprite_selected, const CStr& _textcolor);
 
 	virtual CRect GetListRect() const;
 
 	std::vector<ObjectDef> m_ObjectsDefs;
-	int m_SelectedDef;
+	size_t m_SelectedDef;
 	int m_SelectedColumnOrder;
 
 private:
