+++ source/gui/IGUITextOwner.h
@@ -1,4 +1,4 @@
-/* Copyright (C) 2009 Wildfire Games.
+/* Copyright (C) 2015 Wildfire Games.
  * This file is part of 0 A.D.
  *
  * 0 A.D. is free software: you can redistribute it and/or modify
@@ -35,22 +35,7 @@ GUI Object Base - Text Owner
 #ifndef INCLUDED_IGUITEXTOWNER
 #define INCLUDED_IGUITEXTOWNER
 
-//--------------------------------------------------------
-//  Includes / Compiler directives
-//--------------------------------------------------------
 #include "GUI.h"
-
-//--------------------------------------------------------
-//  Macros
-//--------------------------------------------------------
-
-//--------------------------------------------------------
-//  Types
-//--------------------------------------------------------
-
-//--------------------------------------------------------
-//  Declarations
-//--------------------------------------------------------
 
 /**
  * Framework for handling Output text.
@@ -66,12 +51,12 @@ public:
 	/**
 	 * Adds a text object.
 	 */
-	void AddText(SGUIText * text);
+	void AddText(SGUIText* text);
 
 	/**
 	 * @see IGUIObject#HandleMessage()
 	 */
-	virtual void HandleMessage(SGUIMessage &Message);
+	virtual void HandleMessage(SGUIMessage& Message);
 
 	/**
 	 * @see IGUIObject#UpdateCachedSize()
@@ -88,7 +73,7 @@ public:
 	 * @param clipping Clipping rectangle, don't even add a parameter
 	 *		  to get no clipping.
 	 */
-	virtual void DrawText(int index, const CColor& color, const CPos& pos, float z, const CRect& clipping = CRect());
+	virtual void DrawText(size_t index, const CColor& color, const CPos& pos, float z, const CRect& clipping = CRect());
 
 	/**
 	 * Test if mouse position is over an icon
@@ -100,7 +85,7 @@ protected:
 	/**
 	 * Setup texts. Functions that sets up all texts when changes have been made.
 	 */
-	virtual void SetupText()=0;
+	virtual void SetupText() = 0;
 
 	/**
 	 * Whether the cached text is currently valid (if not then SetupText will be called by Draw)
@@ -115,7 +100,7 @@ protected:
 	/**
 	 * Calculate the position for the text, based on the alignment.
 	 */
-	void CalculateTextPosition(CRect &ObjSize, CPos &TextPos, SGUIText &Text);
+	void CalculateTextPosition(CRect& ObjSize, CPos& TextPos, SGUIText& Text);
 };
 
-#endif
+#endif // INCLUDED_IGUITEXTOWNER
